+++ steamclog/src/main/java/com/steamclock/steamclog/Config.kt
@@ -1,6 +1,5 @@
 package com.steamclock.steamclog
 
-import com.google.firebase.analytics.FirebaseAnalytics
 import java.io.File
 
 /**
@@ -37,19 +36,11 @@ data class Config(
      */
     var requireRedacted: Boolean = false,
 
-    /**
-     * Currently we cannot get the firebase instance from within the Steamclog library, so we
-     * require the calling application to provide it if analytics are desired.
-     */
-    var firebaseAnalytics: FirebaseAnalytics? = null
 ) {
-    constructor(isDebug: Boolean, writeFilePath: File) : this(isDebug, writeFilePath, firebaseAnalytics = null)
-
     override fun toString(): String {
         return "Config(" +
                 "\n  logLevel = $logLevel," +
                 "\n  fileWritePath = $fileWritePath," +
-                "\n  firebaseAnalytics = ${firebaseAnalytics}," +
                 "\n  keepLogsForDays = $keepLogsForDays," +
                 "\n  requireRedacted = $requireRedacted)"
     }